+++ angular-offline.js
@@ -42,6 +42,19 @@ angular
     return this;
   };
 
+
+  /**
+   * Enable or disable always refreshing when online
+   *
+   * @param {boolean} value
+   * @returns {offlineProvider}
+   */
+
+  offlineProvider.alwaysRefresh = function (value) {
+    this._alwaysRefresh = value;
+    return this;
+  };
+
   this.$get = ['$q', '$rootScope', '$window', '$log', 'connectionStatus', '$cacheFactory',
   function ($q, $rootScope, $window, $log, connectionStatus, $cacheFactory) {
     var offline = {
@@ -66,18 +79,20 @@ angular
     }
 
     /**
-     * Clean cache if expired.
+     * Clean cache key
      *
      * @param {object} cache Cache
      * @param {string} key Cache key
      */
 
-    function cleanIfExpired(cache, key) {
+    function clean(cache, key) {
       if (cache === true)
         cache = $requester.defaults.cache || $cacheFactory.get('$http');
       var info = cache.info(key);
-      if (info && info.isExpired)
+
+      if (offlineProvider._alwaysRefresh || (info && info.isExpired)) {
         cache.remove(key);
+      }
     }
 
     /**
@@ -238,8 +253,9 @@ angular
         // For GET method, Angular will handle it.
         if (config.method === 'GET') {
           // Online we clean the cache.
-          if (connectionStatus.isOnline())
-            cleanIfExpired(config.cache, config.url);
+          if (connectionStatus.isOnline()) {
+            clean(config.cache, config.url);
+          }
 
           return config;
         }