+++ changelog.md
@@ -3,4 +3,5 @@
 This project uses Semantic Versioning (2.0).
 
 # Upcoming
+* New feature: Add configuration option to always refresh cache when online
 * Bugfix: Dry out error message strings